+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - service name included in the header
 - refactor name of "Plan" to "Journey"
 - refactor name of "Question" to "Step"
+- refactor redis caching into reusable class
 
 ## [release-002] - 2020-11-16